Полировка и правка мерджа
