+++ main.py
@@ -1,3 +1,5 @@
+from entities import AnimatedSprite
+from our_tools import load_image, all_sprites
 from start_screen import *
 import os
 import pygame
@@ -10,79 +12,14 @@ WIDTH = 600
 HEIGHT = 600
 clock = pygame.time.Clock()
 
-all_sprites = pygame.sprite.Group()
-
-
-def load_image(name, color_key=None):
-    fullname = os.path.join('data', name)
-    try:
-        image = pygame.image.load(fullname)
-    except pygame.error as message:
-        print('Cannot load image:', name)
-        raise SystemExit(message)
-
-    if color_key is not None:
-        if color_key == -1:
-            color_key = image.get_at((0, 0))
-        image.set_colorkey(color_key)
-    else:
-        image = image.convert_alpha()
-    return image
-
-
-class AnimatedSprite(pygame.sprite.Sprite):
-    def __init__(self, sheet, columns, rows, x, y):
-        super().__init__(all_sprites)
-        self.times = 0
-        self.spd = 2
-        self.frames = []
-        self.cut_sheet(sheet, columns, rows)
-        self.cur_frame = 0
-        self.image = self.frames[self.cur_frame]
-        self.rect = self.rect.move(x, y)
-
-    def cut_sheet(self, sheet, columns, rows):
-        self.rect = pygame.Rect(0, 0, sheet.get_width() // columns, sheet.get_height() // rows)
-        for j in range(rows):
-            for i in range(columns):
-                frame_location = (self.rect.w * i, self.rect.h * j)
-                self.frames.append(sheet.subsurface(pygame.Rect(frame_location, self.rect.size)))
-
-    def update(self):
-        if self.times % 6 == 0:
-            self.cur_frame = (self.cur_frame + 1) % len(self.frames)
-            self.image = self.frames[self.cur_frame]
-            self.times = 0
-        self.times += 1
-
-    def move(self, keys):
-        global lr
-        for key in keys:
-            if key == pygame.K_UP:
-                self.rect.y -= self.spd
-            elif key == pygame.K_DOWN:
-                self.rect.y += self.spd
-            elif key == pygame.K_RIGHT:
-                if lr == -1:
-                    lr = 1
-                    self.frames = list(map(lambda x: pygame.transform.flip(x, 1, 0), self.frames))
-                    self.image = self.frames[self.cur_frame]
-                self.rect.x += self.spd
-
-            elif key == pygame.K_LEFT:
-                if lr == 1:
-                    lr = -1
-                    self.frames = list(map(lambda x: pygame.transform.flip(x, 1, 0), self.frames))
-                    self.image = self.frames[self.cur_frame]
-                self.rect.x -= self.spd
-
-
 man = AnimatedSprite(load_image("man_1.png"), 8, 1, 150, 150)
 direct = []
-lr = 1
 running = True
-start_screen()
+username = start_screen()
+score = 5000
 screen = pygame.display.set_mode((WIDTH, HEIGHT))
+level_name = 'level 0'
+pygame.display.set_caption(level_name)
 
 while running:
     for event in pygame.event.get():
@@ -102,7 +39,7 @@ while running:
     pygame.display.flip()
     clock.tick(FPS)
 
-# Вызов экрана конца игры. Пока что только добавляется игрок Vasya.
+# Вызов экрана конца игры. Теперь никнейм берётся из стартскрина, а очки в переменной score.
 pygame.display.set_mode((800, 500))
-game_over('Vasya', 1313)
+game_over(username, score)
 pygame.quit()
