+++ our_tools.py
@@ -1,6 +1,7 @@
 import os
-
 import pygame
+
+all_sprites = pygame.sprite.Group()  # специально группа тут, чтобы не было цикличного импортирования
 
 
 def load_image(name, color_key=None):
